working path to settings, crash fixed
